# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * NyayaSetu AI - Judicial Dashboard & Legal Assistant API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface DashboardSummary {
  totalPendingCases: number;
  totalActiveCases: number;
  totalRegisteredThisWeek: number;
  totalDisposedThisWeek: number;
  totalCourts: number;
  totalJudges: number;
  vacantJudgePositions: number;
  disposalRate: number;
  lastUpdated: string;
}

export interface StateStats {
  stateCode: string;
  stateName: string;
  pendingCases: number;
  activeCases: number;
  registeredThisWeek: number;
  disposedThisWeek: number;
  totalCourts: number;
  disposalRate: number;
}

export interface DistrictStats {
  districtName: string;
  pendingCases: number;
  activeCases: number;
  registeredThisWeek: number;
  disposedThisWeek: number;
}

export interface StateDetails {
  stateCode: string;
  stateName: string;
  pendingCases: number;
  activeCases: number;
  registeredThisWeek: number;
  disposedThisWeek: number;
  totalCourts: number;
  disposalRate: number;
  districts: DistrictStats[];
}

export interface MonthlyTrend {
  month: string;
  filed: number;
  disposed: number;
  pending: number;
}

export interface CourtTypeStats {
  courtType: string;
  pendingCases: number;
  activeCases: number;
  totalCourts: number;
}

export interface CaseCategoryStats {
  category: string;
  count: number;
  percentage: number;
}

export interface ApiError {
  error: string;
}

export interface OpenaiConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface OpenaiMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface OpenaiConversationInput {
  title: string;
}

export interface OpenaiMessageInput {
  content: string;
}

export interface OpenaiConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: OpenaiMessage[];
}

export interface OpenaiError {
  error: string;
}

export type GetMonthlyTrendsParams = {
stateCode?: string;
};

export type GetCaseCategoriesParams = {
stateCode?: string;
};