# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: NyayaSetu AI - Judicial Dashboard & Legal Assistant API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: judicial
    description: Judicial data and case statistics
  - name: openai
    description: OpenAI AI chat operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /judicial/dashboard-summary:
    get:
      operationId: getDashboardSummary
      tags: [judicial]
      summary: Get national dashboard summary stats
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"

  /judicial/states:
    get:
      operationId: listStates
      tags: [judicial]
      summary: List all states with case statistics
      responses:
        "200":
          description: List of states with case data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StateStats"

  /judicial/states/{stateCode}:
    get:
      operationId: getStateDetails
      tags: [judicial]
      summary: Get detailed case stats for a specific state
      parameters:
        - name: stateCode
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: State details with district breakdown
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StateDetails"
        "404":
          description: State not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /judicial/monthly-trends:
    get:
      operationId: getMonthlyTrends
      tags: [judicial]
      summary: Get monthly filing and disposal trends
      parameters:
        - name: stateCode
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Monthly trends
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MonthlyTrend"

  /judicial/court-types:
    get:
      operationId: getCourtTypeBreakdown
      tags: [judicial]
      summary: Get case breakdown by court type
      responses:
        "200":
          description: Court type breakdown
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CourtTypeStats"

  /judicial/top-pending-states:
    get:
      operationId: getTopPendingStates
      tags: [judicial]
      summary: Get top 10 states by pending cases
      responses:
        "200":
          description: Top pending states
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StateStats"

  /judicial/case-categories:
    get:
      operationId: getCaseCategories
      tags: [judicial]
      summary: Get case distribution by category/type
      parameters:
        - name: stateCode
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Case categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CaseCategoryStats"

  /openai/conversations:
    get:
      operationId: listOpenaiConversations
      tags: [openai]
      summary: List all conversations
      responses:
        "200":
          description: List of conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OpenaiConversation"
    post:
      operationId: createOpenaiConversation
      tags: [openai]
      summary: Create a new conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OpenaiConversationInput"
      responses:
        "201":
          description: Created conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiConversation"

  /openai/conversations/{id}:
    get:
      operationId: getOpenaiConversation
      tags: [openai]
      summary: Get conversation with messages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Conversation with messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiConversationWithMessages"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiError"
    delete:
      operationId: deleteOpenaiConversation
      tags: [openai]
      summary: Delete a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiError"

  /openai/conversations/{id}/messages:
    get:
      operationId: listOpenaiMessages
      tags: [openai]
      summary: List messages in a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OpenaiMessage"
    post:
      operationId: sendOpenaiMessage
      tags: [openai]
      summary: Send a text message and receive a streaming text response
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OpenaiMessageInput"
      responses:
        "200":
          description: SSE stream of assistant text chunks
          content:
            text/event-stream: {}

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    DashboardSummary:
      type: object
      properties:
        totalPendingCases:
          type: integer
        totalActiveCases:
          type: integer
        totalRegisteredThisWeek:
          type: integer
        totalDisposedThisWeek:
          type: integer
        totalCourts:
          type: integer
        totalJudges:
          type: integer
        vacantJudgePositions:
          type: integer
        disposalRate:
          type: number
        lastUpdated:
          type: string
          format: date-time
      required:
        - totalPendingCases
        - totalActiveCases
        - totalRegisteredThisWeek
        - totalDisposedThisWeek
        - totalCourts
        - totalJudges
        - vacantJudgePositions
        - disposalRate
        - lastUpdated

    StateStats:
      type: object
      properties:
        stateCode:
          type: string
        stateName:
          type: string
        pendingCases:
          type: integer
        activeCases:
          type: integer
        registeredThisWeek:
          type: integer
        disposedThisWeek:
          type: integer
        totalCourts:
          type: integer
        disposalRate:
          type: number
      required:
        - stateCode
        - stateName
        - pendingCases
        - activeCases
        - registeredThisWeek
        - disposedThisWeek
        - totalCourts
        - disposalRate

    DistrictStats:
      type: object
      properties:
        districtName:
          type: string
        pendingCases:
          type: integer
        activeCases:
          type: integer
        registeredThisWeek:
          type: integer
        disposedThisWeek:
          type: integer
      required:
        - districtName
        - pendingCases
        - activeCases
        - registeredThisWeek
        - disposedThisWeek

    StateDetails:
      type: object
      properties:
        stateCode:
          type: string
        stateName:
          type: string
        pendingCases:
          type: integer
        activeCases:
          type: integer
        registeredThisWeek:
          type: integer
        disposedThisWeek:
          type: integer
        totalCourts:
          type: integer
        disposalRate:
          type: number
        districts:
          type: array
          items:
            $ref: "#/components/schemas/DistrictStats"
      required:
        - stateCode
        - stateName
        - pendingCases
        - activeCases
        - registeredThisWeek
        - disposedThisWeek
        - totalCourts
        - disposalRate
        - districts

    MonthlyTrend:
      type: object
      properties:
        month:
          type: string
        filed:
          type: integer
        disposed:
          type: integer
        pending:
          type: integer
      required:
        - month
        - filed
        - disposed
        - pending

    CourtTypeStats:
      type: object
      properties:
        courtType:
          type: string
        pendingCases:
          type: integer
        activeCases:
          type: integer
        totalCourts:
          type: integer
      required:
        - courtType
        - pendingCases
        - activeCases
        - totalCourts

    CaseCategoryStats:
      type: object
      properties:
        category:
          type: string
        count:
          type: integer
        percentage:
          type: number
      required:
        - category
        - count
        - percentage

    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    OpenaiConversation:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - createdAt

    OpenaiMessage:
      type: object
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        role:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - conversationId
        - role
        - content
        - createdAt

    OpenaiConversationInput:
      type: object
      properties:
        title:
          type: string
      required:
        - title

    OpenaiMessageInput:
      type: object
      properties:
        content:
          type: string
      required:
        - content

    OpenaiConversationWithMessages:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
        messages:
          type: array
          items:
            $ref: "#/components/schemas/OpenaiMessage"
      required:
        - id
        - title
        - createdAt
        - messages

    OpenaiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error