# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * NyayaSetu AI - Judicial Dashboard & Legal Assistant API
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardSummary {
  totalPendingCases: number;
  totalActiveCases: number;
  totalRegisteredThisWeek: number;
  totalDisposedThisWeek: number;
  totalCourts: number;
  totalJudges: number;
  vacantJudgePositions: number;
  disposalRate: number;
  lastUpdated: Date;
}