# lib/api-zod/src/generated/types/districtStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * NyayaSetu AI - Judicial Dashboard & Legal Assistant API
 * OpenAPI spec version: 0.1.0
 */

export interface DistrictStats {
  districtName: string;
  pendingCases: number;
  activeCases: number;
  registeredThisWeek: number;
  disposedThisWeek: number;
}