# lib/api-zod/src/generated/types/stateDetails.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * NyayaSetu AI - Judicial Dashboard & Legal Assistant API
 * OpenAPI spec version: 0.1.0
 */
import type { DistrictStats } from './districtStats';

export interface StateDetails {
  stateCode: string;
  stateName: string;
  pendingCases: number;
  activeCases: number;
  registeredThisWeek: number;
  disposedThisWeek: number;
  totalCourts: number;
  disposalRate: number;
  districts: DistrictStats[];
}